Write tests for and implement the Ship factory fn
More specifically, the methods within the Ship factory that were tested
are hit() which should increase the number of 'hits' on the ship it was
called on.

isSunk() is the other method that should return true when the number of
hits on the ship are equal to the length of the ship and false otherwise.