Setup Ship factory for Gameboard testing
The Ship factory needed a way to store the coordinates a ship will
occupy on the board.

The Ship factory also needed a way to specify whether the a ship will
will be oriented vertically or horizontally.

Checking for passing valid arguments in was also implemented into the
Ship factory.