Add an opponent property to Player objects
An opponent property was added to Player objects in order for the
calculateDecision method to make sure it doesn't choose a cell on the
opponent's gameboard that was already attacked.

The game loop was also fixed so the turn doesn't switch over to the
CPU when the user clicks on a cell that was already attacked. The text
announcer also alerts the user if they made this mistake.

The text announcer now displays whether or not the CPU player's attack
hit or missed.